added a couple routes. began creating post page

+++ controllers/api/postRoutes.js
@@ -54,14 +54,39 @@ router.post("/", async (req, res) => {
 
 router.delete("/:id", async (req, res) => {
   try {
-    const postData = await Post.destroy({
+    if (!req.session.logged_in) {
+      return res
+        .status(401)
+        .json({ message: "Unauthorized: User not logged in" });
+    }
+
+    const userId = req.session.user_id;
+    const post = await Post.findOne({
+      where: {
+        id: req.params.id,
+        created_by: userId,
+      },
+    });
+
+    if (!post) {
+      return res.status(404).json({
+        message: "Unauthorized: Incorrect User or Post does not exist",
+      });
+    }
+
+    const deletedRows = await Post.destroy({
       where: {
         id: req.params.id,
       },
     });
-    res.status(404).json({ message: "Post does not exist" });
+
+    if (deletedRows === 0) {
+      return res.status(404).json({ message: "Post does not exist" });
+    }
+
+    return res.status(200).json({ message: "Post deleted successfully!" });
   } catch (err) {
-    console.err(err);
-    res.status(500).json(err);
+    console.error(err);
+    return res.status(500).json({ error: "Internal Server Error" });
   }
 });
